Carry a kernel interface index at its real width
An OS interface index is a 32-bit unsigned number that, on Linux, comes
from a counter which only ever increases - a long-lived host that churns
veth or container interfaces will issue values well past 65535. PMIx
carried one in a uint16_t field and handed it back through helpers that
did not agree on a width:

  pmix_pif_t.if_kernel_index      uint16_t
  pmix_ifnametokindex()           int16_t
  pmix_ifaddrtokindex()           int16_t
  pmix_ifindextokindex()          int
  pmix_ifkindextoname()           int parameter
  pmix_ifkindextoaddr()           int parameter
  pmix_ifmatches()                int parameter

The disagreement bites well before the field overflows. An index of
40000 stores in the uint16_t perfectly well, and pmix_ifindextokindex()
returns it correctly - but pmix_ifnametokindex() returns it as -25536,
and every caller of these helpers screens the result with "if (0 >
kindex)". So the same interface was simultaneously findable by list
index and missing by name, which is a silent wrong answer rather than an
error: an if_include entry naming that interface is quietly skipped.

Everything now uses one width. The field is a uint32_t, both int16_t
returns become int - which is what every consumer of a kernel index
already took, and what PRRTE's own kindex variables already are - and
the constructor's unassigned sentinel becomes UINT32_MAX.

Widening the field made the compiler point out four comparisons that
were relying on the old integer promotion, and they are worth more than
a cast: pmix_ifkindextoname(), pmix_ifkindextoaddr() and pmix_ifmatches()
all take a signed parameter that callers legitimately feed a -1 from a
failed lookup, and converting that to unsigned would have produced
UINT32_MAX - the exact value the constructor leaves in an entry no
component has filled in. Each now rejects a negative index up front.

test/unit/util/util_if.c builds its interface list by hand, which is the
only way to exercise this: no developer's machine has an interface index
above 32767. Both halves were verified by re-breaking - restoring the
int16_t return fails three of the new cases (nametokindex disagreeing
with indextokindex at 40000), and restoring the uint16_t field fails the
70000 case on its own.

Note for anything building against these headers: pmix_pif_t is in the
installed src/util/pmix_if.h and PRRTE's oob does walk pmix_if_list and
construct pmix_pif_t objects directly, so it needs a rebuild against
this. It needs no source change - its three uses of if_kernel_index all
go through an int already - and it calls neither of the two functions
whose return type changed.

+++ src/util/pmix_if.h
@@ -58,7 +58,7 @@ typedef struct pmix_pif_t {
     pmix_list_item_t super;
     char if_name[PMIX_IF_NAMESIZE + 1];
     int if_index;
-    uint16_t if_kernel_index;
+    uint32_t if_kernel_index;
     uint16_t af_family;
     int if_flags;
     int if_speed;
@@ -101,14 +101,14 @@ PMIX_EXPORT int pmix_ifnametoindex(const char *if_name);
  *  @param if_name (IN)  Interface name
  *  @return              Interface kernel index
  */
-PMIX_EXPORT int16_t pmix_ifnametokindex(const char *if_name);
+PMIX_EXPORT int pmix_ifnametokindex(const char *if_name);
 
 /*
  *  Attempt to resolve an address (given as either IPv4/IPv6 string
  *  or hostname) and return the kernel index of the interface
  *  that is on the same network as the specified address
  */
-PMIX_EXPORT int16_t pmix_ifaddrtokindex(const char *if_addr);
+PMIX_EXPORT int pmix_ifaddrtokindex(const char *if_addr);
 
 /**
  *  Lookup an interface by pmix_list index and return its kernel index.
